% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_path.R
\name{retrieve_path}
\alias{retrieve_path}
\title{Retrieve and escape a path from the clipboard}
\usage{
retrieve_path(wrap_quotes = TRUE)
}
\arguments{
\item{wrap_quotes}{if TRUE, wrap in double quotes (character vector) if necessary}
}
\value{
a character vector or NULL
}
\description{
#' All backslashes are converted to slashes, and the path is quoted if necessary.
}
