% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_path.R
\name{paste_path}
\alias{paste_path}
\title{Paste a path in the current cursor position.
Paste a path in the current cursor position.
All backslashes are converted to slashes, and the path is quoted if necessary.}
\usage{
paste_path()
}
\value{
nothing
}
\description{
Paste a path in the current cursor position.
Paste a path in the current cursor position.
All backslashes are converted to slashes, and the path is quoted if necessary.
}
\examples{

paste_path()

}
