% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_path.R
\name{paste_path}
\alias{paste_path}
\title{Paste a path in the current cursor position.}
\usage{
paste_path(wrap_quotes = TRUE)
}
\arguments{
\item{wrap_quotes}{if TRUE, wrap in double quotes (character vector) if necessary}
}
\value{
nothing
}
\description{
All backslashes are converted to slashes, and the path is quoted if necessary.
}
\examples{

paste_path()

}
