% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_path.R
\name{retrieve_path}
\alias{retrieve_path}
\title{Retrieve and escape a path from the clipboard}
\usage{
retrieve_path(wrap_quotes = TRUE)
}
\arguments{
\item{wrap_quotes}{if TRUE, wrap in a character vector if not necessary}
}
\value{
a character vector or NULL
}
\description{
Retrieve and escape a path from the clipboard
}
